pyproject.toml
[tool.poetry]
name = "service_dataset"
version = "0.3.0"
description = ""
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "0.62.0"
fastapi-utils = "0.2.1"
uvicorn = "0.12.3"
gunicorn = "20.0.4"
uvloop = "0.14.0"
httptools = "0.1.1"
pyjwt = "1.4.2"
python-multipart = "0.0.5"
python-json-logger = "0.1.11"
aiofiles = "0.6.0"
redis = "3.5.3"
minio = "7.0.3"
xmltodict = "0.12.0"
pika = "1.1.0"
fastapi-sqlalchemy = "0.2.1"
psycopg2 = "2.8.6"
bids-validator = "1.7.3"
python-dotenv = "0.19.1"
httpx = "^0.22.0"
pytest = "^7.1.1"
pilot-platform-common = "^0.0.25"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation = "^0.30b1"
opentelemetry-instrumentation-httpx = "^0.30b1"
opentelemetry-instrumentation-asyncpg = "^0.30b1"
opentelemetry-instrumentation-sqlalchemy = "^0.30b1"
opentelemetry-instrumentation-fastapi = "^0.30b1"


[tool.poetry.dev-dependencies]
pytest-env = "0.6.2"
requests = "^2.27.1"
flake8 = "^4.0.1"
pytest-httpx = "^0.20.0"
pytest-asyncio = "^0.18.2"
async-asgi-testclient = "^1.4.10"
pytest-cov = "^3.0.0"
testcontainers = "3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
